decenthrathon/requirements.txt
# 🏆 Ultra-Optimized Customer Segmentation Pipeline
# Core Dependencies with Fixed Versions for Reproducibility

# Data Processing
pandas==2.1.3
numpy==1.24.3
pyarrow==14.0.1

# Machine Learning
scikit-learn==1.3.2
hdbscan==0.8.33
umap-learn==0.5.4

# Visualization (Optional)
matplotlib==3.8.2
seaborn==0.13.0

# AI/ML Enhancement
openai==1.3.7

# Configuration
python-dotenv==1.0.0

# Parallel Processing
joblib==1.3.2
tqdm==4.66.1

# Development (Optional)
pytest==7.4.3
black==23.11.0 